style: 폴더구조 및 import문 개선
