+++ src/app/infocard/(form)/2/page.tsx
@@ -3,20 +3,11 @@
 import Link from 'next/link';
 import { useRef } from 'react';
 
-import {
-  bgSub,
-  button,
-  flexCol,
-  flexColCenter,
-  flexRow,
-  optionalText,
-  subtitleText,
-  titleLg,
-  titleMd,
-  whiteText,
-} from '@/styles/ogoo';
-import { input } from '@/styles/ogoo/input.css';
-import { cn } from '@/styles/utils';
+import { button, input } from '@/styles/ogoo';
+import { flexCol, flexColCenter, flexRow } from '@/styles/ogoo/alignment.css';
+import { bgSub, optionalText, subtitleText, whiteText } from '@/styles/ogoo/colors.css';
+import { titleLg, titleMd } from '@/styles/ogoo/typography.css';
+import { cn } from '@/utils';
 
 export default function InfoCardProcessTwo() {
   const fileRef = useRef<HTMLInputElement>(null);
