+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './auth';
+export * from './common';
+export * from './petcard';
