+++ src/components/petcard/index.ts
@@ -1,0 +1,2 @@
+export * from './progressbar';
+export * from './tab';
