+++ src/styles/ogoo/colors.css.ts
@@ -1,0 +1,30 @@
+import { sprinkles } from '..';
+
+export const primary = sprinkles({ color: 'primary' });
+export const secondary = sprinkles({ color: 'secondary' });
+export const subText = sprinkles({ color: 'subText' });
+export const optionalText = sprinkles({ color: 'optionalText' });
+export const subtitleText = sprinkles({ color: 'subtitleText' });
+
+export const whiteText = sprinkles({ color: 'whiteText' });
+export const blackText = sprinkles({ color: 'blackText' });
+
+export const bgPrimary = sprinkles({ background: 'primary' });
+export const bgSubPrimary = sprinkles({ background: 'primarySub' });
+export const bgSecondary = sprinkles({ background: 'secondary' });
+export const bgSub = sprinkles({ background: 'sub' });
+export const bgOptional = sprinkles({ background: 'optional' });
+export const bgKakao = sprinkles({ background: 'kakao' });
+
+export const inputColor = sprinkles({
+  color: 'input',
+});
+export const dividerColor = sprinkles({
+  color: 'divider',
+});
+export const inputBorder = sprinkles({
+  borderColor: 'input',
+});
+export const dividerBorder = sprinkles({
+  borderColor: 'divider',
+});
