+++ src/styles/ogoo/container.css.ts
@@ -9,11 +9,3 @@ export const box = sprinkles({
   color: 'subText',
   background: 'optional',
 });
-
-export const inputBorderColor = sprinkles({
-  borderColor: 'input',
-});
-
-export const dividerColor = sprinkles({
-  borderColor: 'divider',
-});
