+++ src/styles/ogoo/index.ts
@@ -1,4 +1,2 @@
-export * from './alignment.css';
 export * from './button.css';
-export * from './container.css';
-export * from './typography.css';
+export * from './input.css';
