+++ src/styles/ogoo/input.css.ts
@@ -9,24 +9,23 @@ export const input = recipe({
 
   variants: {
     color: {
-      inputColor: themeVars.borderColors.input,
+      inputColor: themeVars.colors.input,
     },
-    size: {
+    line: {
       single: {
         width: '100%',
         height: 60,
         borderRadius: 12,
         fontSize: 18,
         fontWeight: 500,
-        border: `solid ${themeVars.borderColors.input}`,
+        border: `solid ${themeVars.colors.input}`,
         padding: 20,
       },
     },
   },
 
-  // fallback으로 설정하는 variant
   defaultVariants: {
     color: 'inputColor',
-    size: 'single',
+    line: 'single',
   },
 });
